+++ .gitignore
@@ -14,3 +14,4 @@ node_modules/
 # Папка с собранными файлами проекта
 build/
 public/js
+
