+++ package.json
@@ -12,6 +12,7 @@
   "devDependencies": {
     "eslint": "5.13.0",
     "eslint-config-htmlacademy": "0.5.2",
+    "html-webpack-plugin": "3.2.0",
     "webpack": "4.29.5",
     "webpack-cli": "3.2.3",
     "webpack-dev-server": "3.2.0"
